visaphoto_app/requirements.txt
opencv-python
rembg
Pillow
face_recognition
